moved App.bridge into its own bridge.js file. Installed and set up jest
