+++ webpack.config.js
@@ -30,4 +30,11 @@ module.exports = {
       },
     }),
   ],
+  resolve: {
+    extensions: ['.js', '.jsx'],
+    modules: [
+      'node_modules',
+      '/src/tests',
+    ],
+  },
 };
